Merge branch 'master' of https://github.com/msimonides/homerplayer
Rework handling of settings access to work well with "big buttons" yet
remain hard to access accidentally. Default is just an ordinary settings
button (for initial setup) with two options to make it difficult to
accidentally get into settings. Refactored the logic for this to a
common place so additonal modes are easy to add if needed.

Required reworking two screens for the "gear" icon, and those were
reworked to also scale from the smallest to largest phone screens
on the emulator with no significant difference.